fix misleading paragraph in readme
The old paragraph contained a dead link as was noted [Github issue #3089](https://github.com/rakudo/rakudo/issues/3089); Moreover the paragraph reflected when there were multiple partial implementations of Perl 6. As Rokudo is now the most complete implementation, it is better to reflect that. Plus, the page comparing the compilers seems to be gone from the website, or at least I couldn't find it. Thus, I substituted for a link to the FAQ explaining the different perl 6 vs roku vs rokudo.